Ändring av StartView och CreateView till tomma
