+++ package.json
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.1.4",
-    "concurrently": "^9.0.1",
+    "concurrently": "^9.1.0",
     "eslint": "^9.13.0",
     "eslint-plugin-vue": "^9.29.0",
     "nodemon": "^3.1.7",
